# openapi.yaml
openapi: 3.0.0
info:
  title: Mock User Management API
  version: 1.0.0
  description: API for testing telemetry libraries with varied endpoint behaviors.
paths:
  /:
    get:
      summary: Root endpoint to check API status
      responses:
        '200':
          description: API is running.
  /users:
    get:
      summary: List all users
      responses:
        '200':
          description: A list of users.
    post:
      summary: Create a new user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
      responses:
        '201':
          description: User created.
  /users/{id}:
    get:
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A user object.
        '404':
          description: User not found.
    put:
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
      responses:
        '200':
          description: User updated.
        '404':
          description: User not found.
    delete:
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User deleted.
        '404':
          description: User not found.
  /error:
    get:
      summary: Endpoint that always fails
      responses:
        '500':
          description: Internal Server Error.
  /slow:
    get:
      summary: Endpoint that is always slow
      responses:
        '200':
          description: Success after delay.
  /fast:
    get:
      summary: Endpoint that is always fast
      responses:
        '200':
          description: Success.
  /random-error:
    get:
      summary: Endpoint that fails randomly
      responses:
        '200':
          description: Success.
        '500':
          description: Random failure.